test: 添加异步导出单测，将导出单测分离成两个
